Fix for OUJS
OUJS has made a change recently to require SPDX codes for OSI approved licensing.

In order to improve the appearance of your script homepages it would be appreciated if you could modify all of your affected scripts. This appears to be the two on the site... but you do have some requires elsewhere that may need some addressing.

You also seem to be out of sync with `// @name        MouseHunt AutoBot ENHANCED + REVAMP`. A reimport should fix things if you are utilizing the webhook.

Until this change is made you will be unable to update those affected scripts.

Thanks,
OUJS Staff